refactor(react-query): expose provider entries
