+++ packages/suite/package.json
@@ -107,6 +107,7 @@
         "@suite/modal": "workspace:*",
         "@suite/nfc": "workspace:*",
         "@suite/onboarding-components": "workspace:*",
+        "@suite/react-query": "workspace:*",
         "@suite/receive": "workspace:*",
         "@suite/recovery": "workspace:*",
         "@suite/router": "workspace:*",
@@ -120,7 +121,6 @@
         "@suite/tx-simulation": "workspace:^",
         "@suite/ui-animations": "workspace:*",
         "@suite/wallet": "workspace:*",
-        "@tanstack/react-query": "5.101.2",
         "@testing-library/jest-dom": "6.9.1",
         "@trezor/address-validator": "workspace:*",
         "@trezor/analytics-uploader": "workspace:*",
