+++ packages/suite/src/hooks/wallet/allowance/useAllowanceCompose.ts
@@ -2,8 +2,8 @@ import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useForm } from 'react-hook-form';
 
 import { isFulfilled } from '@reduxjs/toolkit';
-import { useMutation } from '@tanstack/react-query';
 
+import { useMutation } from '@suite-common/react-query';
 import { DEFAULT_PAYMENT, DEFAULT_VALUES } from '@suite-common/wallet-constants';
 import {
     type ComposeAllowanceTransactionThunkParams,
