--- suite-common/react-query/src/config/index.ts
@@ -1,1 +1,0 @@
-export const isDevEnv = process.env.NODE_ENV !== 'production';
