--- suite-common/react-query/src/constants/mutationKeys.ts
@@ -1,3 +1,0 @@
-import { type AllowedMutationKey } from '../types';
-
-export const desktopMutationKeys = {} as const satisfies Record<string, AllowedMutationKey>;
