+++ suite-common/react-query/src/index.ts
@@ -1,6 +1,8 @@
 export {
+    MutationCache,
     QueryClient,
     QueryClientProvider,
+    QueryCache,
     useQuery,
     useInfiniteQuery,
     useMutation,
@@ -12,5 +14,4 @@ export {
     keepPreviousData,
 } from '@tanstack/react-query';
 export * from './constants/queryKeys';
-export * from './constants/mutationKeys';
-// QueryClientProvider wrappers are not exported here, to keep this package compatible with nodeJS-only environments (which can't parse .tsx)
+export type { AllowedQueryKey } from './types';
