+++ suite-common/react-query/src/types/index.ts
@@ -1,9 +1,5 @@
-import type { MutationKey, QueryKey } from '@tanstack/react-query';
+import type { QueryKey } from '@tanstack/react-query';
 
 type AllQueryKey = QueryKey | QueryKey[];
 
 export type AllowedQueryKey = AllQueryKey | ((...args: any[]) => AllQueryKey);
-
-type AllMutationKey = MutationKey | MutationKey[];
-
-export type AllowedMutationKey = AllMutationKey | ((...args: any[]) => AllMutationKey);
