+++ suite-native/app/package.json
@@ -45,7 +45,6 @@
         "@suite-common/discreet-mode": "workspace:*",
         "@suite-common/firmware-authenticity": "workspace:*",
         "@suite-common/formatters": "workspace:*",
-        "@suite-common/react-query": "workspace:^",
         "@suite-common/suite-constants": "workspace:*",
         "@suite-common/suite-sync": "workspace:*",
         "@suite-common/suite-types": "workspace:*",
@@ -89,6 +88,7 @@
         "@suite-native/module-trading": "workspace:*",
         "@suite-native/module-transactions": "workspace:*",
         "@suite-native/navigation": "workspace:*",
+        "@suite-native/react-query": "workspace:*",
         "@suite-native/sentry": "workspace:*",
         "@suite-native/services": "workspace:*",
         "@suite-native/settings": "workspace:*",
