+++ suite-native/react-query/tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "extends": "../../tsconfig.base.json",
+    "compilerOptions": { "outDir": "libDev" },
+    "references": [
+        {
+            "path": "../../suite-common/react-query"
+        },
+        {
+            "path": "../../suite-common/suite-utils"
+        }
+    ]
+}
