+++ suite/react-query/src/index.ts
@@ -1,0 +1,2 @@
+export { ReactQueryProvider } from './ReactQueryProvider';
+export { desktopQueryKeys } from './constants/queryKeys';
